<SearchResultNext /> -> <SearchResult />, with deprecation warning.
